Fix light computing
Fix light computing bug with Light Pumpkin, lit furnaces, and others.
