+++ src/pocketmine/block/Block.php
@@ -320,16 +320,14 @@ class Block extends Position implements BlockIds, Metadatable{
 					self::$hardness[$id] = $block->getHardness();
 					self::$light[$id] = $block->getLightLevel();
 
-					if($block->isSolid()){
-						if($block->isTransparent()){
-							if($block instanceof Lava){
-								self::$lightFilter[$id] = 15;
-							}elseif($block instanceof Liquid or $block instanceof Ice){
-								self::$lightFilter[$id] = 2;
-							}else{
-								self::$lightFilter[$id] = 1;
-							}
-						}elseif($block->getId() == Block::GLOWSTONE){
+					if($block->isTransparent()){
+						if($block instanceof Water or $block instanceof Ice){
+							self::$lightFilter[$id] = 2;
+						}else{
+							self::$lightFilter[$id] = 1;
+						}
+					}elseif($block->isSolid()){
+						if($block instanceof SolidLight){
 							self::$lightFilter[$id] = 1;
 						}else{
 							self::$lightFilter[$id] = 15;
