+++ src/pocketmine/block/EndPortalFrame.php
@@ -24,7 +24,7 @@ namespace pocketmine\block;
 use pocketmine\item\Item;
 use pocketmine\math\AxisAlignedBB;
 
-class EndPortalFrame extends Solid{
+class EndPortalFrame extends Solid implements SolidLight{
 
 	protected $id = self::END_PORTAL_FRAME;
