+++ src/pocketmine/block/GlowingObsidian.php
@@ -22,7 +22,7 @@
 namespace pocketmine\block;
 
 
-class GlowingObsidian extends Transparent{
+class GlowingObsidian extends Solid implements SolidLight{
 
 	protected $id = self::GLOWING_OBSIDIAN;
