+++ src/pocketmine/block/GlowingRedstoneOre.php
@@ -24,7 +24,7 @@ namespace pocketmine\block;
 use pocketmine\item\Item;
 use pocketmine\level\Level;
 
-class GlowingRedstoneOre extends RedstoneOre{
+class GlowingRedstoneOre extends RedstoneOre implements SolidLight{
 
 	protected $id = self::GLOWING_REDSTONE_ORE;
